+++ source/shared_lib/sources/util/conversion.cpp
@@ -218,9 +218,9 @@ namespace Shared {
 		double getTimeDuationMinutes(int frames, int updateFps) {
 			int framesleft = frames;
 			double hours = (int) ((int) frames / (float) updateFps / 3600.0f);
-			framesleft = framesleft - hours * 3600 * updateFps;
+			framesleft = (int) (framesleft - hours * 3600 * updateFps);
 			double minutes = (int) ((int) framesleft / (float) updateFps / 60.0f);
-			framesleft = framesleft - minutes * 60 * updateFps;
+			framesleft = (int) (framesleft - minutes * 60 * updateFps);
 			double seconds = (int) ((int) framesleft / (float) updateFps);
 
 			double result = (hours * 60.0) + minutes;
